requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Werkzeug>=3.0.3
sentence-transformers>=2.7.0
scikit-learn>=1.4.0
opencv-python>=4.9.0.80
fpdf2>=2.7.9
Pillow>=10.2.0
numpy>=1.26.0
pandas>=2.2.0
